存档

+++ software/include/xcustom64.h
@@ -48,9 +48,9 @@
 // rd, rs1, and rs2 are data
 // rd_n, rs_1, and rs2_n are the register numbers to use
 #define ROCC_INSTRUCTION_R_R_R(X, rd, rs1, rs2, funct, rd_n, rs1_n, rs2_n) { \
-    register uint32_t rd_  asm ("x" # rd_n);                                 \
-    register uint32_t rs1_ asm ("x" # rs1_n) = (uint32_t) rs1;               \
-    register uint32_t rs2_ asm ("x" # rs2_n) = (uint32_t) rs2;               \
+    register uint64_t rd_  asm ("x" # rd_n);                                 \
+    register uint64_t rs1_ asm ("x" # rs1_n) = (uint64_t) rs1;               \
+    register uint64_t rs2_ asm ("x" # rs2_n) = (uint64_t) rs2;               \
     asm volatile (                                                           \
         ".word " STR(CUSTOMX(X, 1, 1, 1, rd_n, rs1_n, rs2_n, funct)) "\n\t"  \
         : "=r" (rd_)                                                         \
@@ -59,8 +59,8 @@
   }
 
 #define ROCC_INSTRUCTION_R_R_I(X, rd, rs1, rs2, funct, rd_n, rs1_n) {     \
-    register uint32_t rd_  asm ("x" # rd_n);                              \
-    register uint32_t rs1_ asm ("x" # rs1_n) = (uint32_t) rs1;            \
+    register uint64_t rd_  asm ("x" # rd_n);                              \
+    register uint64_t rs1_ asm ("x" # rs1_n) = (uint64_t) rs1;            \
     asm volatile (                                                        \
         ".word " STR(CUSTOMX(X, 1, 1, 0, rd_n, rs1_n, rs2, funct)) "\n\t" \
         : "=r" (rd_) : [_rs1] "r" (rs1_));                                \
@@ -68,7 +68,7 @@
   }
 
 #define ROCC_INSTRUCTION_R_I_I(X, rd, rs1, rs2, funct, rd_n) {           \
-    register uint32_t rd_  asm ("x" # rd_n);                             \
+    register uint64_t rd_  asm ("x" # rd_n);                             \
     asm volatile (                                                       \
         ".word " STR(CUSTOMX(X, 1, 0, 0, rd_n, rs1, rs2, funct)) "\n\t"  \
         : "=r" (rd_));                                                   \
@@ -76,15 +76,15 @@
   }
 
 #define ROCC_INSTRUCTION_I_R_R(X, rd, rs1, rs2, funct, rs1_n, rs2_n) {    \
-    register uint32_t rs1_ asm ("x" # rs1_n) = (uint32_t) rs1;            \
-    register uint32_t rs2_ asm ("x" # rs2_n) = (uint32_t) rs2;            \
+    register uint64_t rs1_ asm ("x" # rs1_n) = (uint64_t) rs1;            \
+    register uint64_t rs2_ asm ("x" # rs2_n) = (uint64_t) rs2;            \
     asm volatile (                                                        \
         ".word " STR(CUSTOMX(X, 0, 1, 1, rd, rs1_n, rs2_n, funct)) "\n\t" \
         :: [_rs1] "r" (rs1_), [_rs2] "r" (rs2_));                         \
   }
 
 #define ROCC_INSTRUCTION_I_R_I(X, rd, rs1, rs2, funct, rs1_n) {         \
-    register uint32_t rs1_ asm ("x" # rs1_n) = (uint32_t) rs1;          \
+    register uint64_t rs1_ asm ("x" # rs1_n) = (uint64_t) rs1;          \
     asm volatile (                                                      \
         ".word " STR(CUSTOMX(X, 0, 1, 0, rd, rs1_n, rs2, funct)) "\n\t" \
         :: [_rs1] "r" (rs1_));                                          \
